pyproject.toml
[tool.poetry]
name = "ai-text-demo"
version = "0.1.0"
description = ""
authors = ["Petr Janik <peta.janik@email.cz>"]
readme = "README.md"
packages = [{include = "ai_text_demo"}]

[tool.poetry.dependencies]
python = "^3.9"
weaviate-client = "^3.23.2"
python-dotenv = "^1.0.0"
pandas = "^2.1.0"
unstructured = {extras = ["pdf"], version = "^0.10.11"}
openpyxl = "^3.1.2"
langchain = "^0.0.279"
openai = "^0.28.0"
pypdf = "^3.15.4"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
tiktoken = "^0.4.0"
pymilvus = "^2.3.0"
towhee = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
